# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** e.g. ENEM, Medicina, Direito, Concurso Federal */
  goal: string;
  /** @nullable */
  school?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  goal?: string;
  /** @nullable */
  school?: string | null;
}

export interface GamificationStats {
  xp: number;
  level: number;
  xpToNextLevel: number;
  streak: number;
  coins: number;
  rank: number;
  totalStudyMinutes: number;
  medalsEarned: number;
}

export interface Medal {
  id: number;
  name: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
  /** common, rare, epic, legendary */
  rarity: string;
}

export interface DashboardSummary {
  studyMinutesToday: number;
  questionsAnsweredToday: number;
  simuladosCompleted: number;
  flashcardsReviewedToday: number;
  currentStreak: number;
  xpThisWeek: number;
  nextMilestone: string;
}

export interface PerformancePoint {
  date: string;
  xpEarned: number;
  minutesStudied: number;
  questionsCorrect: number;
  questionsTotal: number;
}

export interface SubjectScore {
  subject: string;
  score: number;
  totalQuestions: number;
  correctAnswers: number;
  /** up, down, stable */
  trend: string;
}

export interface StudyPlan {
  id: number;
  title: string;
  goal: string;
  /** @nullable */
  targetDate?: string | null;
  subjects: string[];
  hoursPerDay: number;
  progress: number;
  createdAt: string;
}

export interface StudyPlanInput {
  title: string;
  goal: string;
  /** @nullable */
  targetDate?: string | null;
  subjects: string[];
  hoursPerDay: number;
}

export interface StudyPlanUpdate {
  title?: string;
  goal?: string;
  /** @nullable */
  targetDate?: string | null;
  subjects?: string[];
  hoursPerDay?: number;
}

export interface StudySession {
  id: number;
  planId: number;
  subject: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  xpEarned: number;
  createdAt: string;
}

export interface StudySessionInput {
  subject: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
}

export interface FlashcardDeck {
  id: number;
  title: string;
  subject: string;
  cardCount: number;
  dueCount: number;
  masteredCount: number;
  createdAt: string;
}

export interface FlashcardDeckInput {
  title: string;
  subject: string;
}

export interface Flashcard {
  id: number;
  deckId: number;
  front: string;
  back: string;
  easeFactor: number;
  interval: number;
  repetitions: number;
  /** @nullable */
  nextReviewAt?: string | null;
  mastered: boolean;
}

export interface FlashcardInput {
  front: string;
  back: string;
}

export interface FlashcardReview {
  /** 0-5 SM2 quality rating */
  quality: number;
}

export interface Simulado {
  id: number;
  title: string;
  /** ENEM, FUVEST, Concurso, Custom */
  type: string;
  /** @nullable */
  subject?: string | null;
  questionCount: number;
  durationMinutes: number;
  /** facil, medio, dificil */
  difficulty: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  score?: number | null;
  createdAt: string;
}

export interface SimuladoInput {
  title: string;
  type: string;
  /** @nullable */
  subject?: string | null;
  difficulty: string;
}

export interface Alternative {
  id: string;
  text: string;
}

export interface Question {
  id: number;
  text: string;
  subject: string;
  alternatives: Alternative[];
}

export interface SimuladoDetail {
  id: number;
  title: string;
  type: string;
  difficulty: string;
  durationMinutes: number;
  questions: Question[];
}

export interface AnswerItem {
  questionId: number;
  selectedAlternative: string;
}

export interface SimuladoAnswers {
  answers: AnswerItem[];
}

export interface SimuladoResult {
  id: number;
  simuladoId: number;
  simuladoTitle: string;
  score: number;
  correctCount: number;
  totalCount: number;
  xpEarned: number;
  timeSpentMinutes: number;
  completedAt: string;
  subjectBreakdown: SubjectScore[];
}

export interface CompetenciaScore {
  number: number;
  description: string;
  score: number;
  maxScore: number;
  feedback: string;
}

export interface Redacao {
  id: number;
  theme: string;
  content: string;
  /** @nullable */
  score?: number | null;
  maxScore: number;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  competencias?: CompetenciaScore[] | null;
  createdAt: string;
}

export interface RedacaoInput {
  theme: string;
  content: string;
}

export interface RankingEntry {
  rank: number;
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  xp: number;
  level: number;
  streak: number;
  goal: string;
  isCurrentUser: boolean;
}

export interface Mission {
  id: number;
  title: string;
  description: string;
  icon: string;
  xpReward: number;
  coinReward: number;
  completed: boolean;
  progress: number;
  target: number;
  /** study, flashcard, simulado, redacao */
  type: string;
}