# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AprovaJá API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flashcards:
    get:
      operationId: getFlashcards
      tags: [flashcards]
      summary: List flashcard decks
      responses:
        "200":
          description: Flashcard decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardDeck"
    post:
      operationId: createFlashcardDeck
      tags: [flashcards]
      summary: Create a flashcard deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardDeckInput"
      responses:
        "201":
          description: Created deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
  /flashcards/{deckId}/cards:
    get:
      operationId: getFlashcardCards
      tags: [flashcards]
      summary: Get cards in a deck
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
    post:
      operationId: createFlashcard
      tags: [flashcards]
      summary: Add a card to a deck
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardInput"
      responses:
        "201":
          description: Created flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
  /flashcards/{deckId}/cards/{cardId}/review:
    post:
      operationId: reviewFlashcard
      tags: [flashcards]
      summary: Submit a flashcard review result
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: integer
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardReview"
      responses:
        "200":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"

  # ── Simulados ──────────────────────────────────────────────────────────────
...
  /redacoes:
    get:
      operationId: getRedacoes
      tags: [redacao]
      summary: List user essays
      responses:
        "200":
          description: Essays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Redacao"
    post:
      operationId: submitRedacao
      tags: [redacao]
      summary: Submit an essay for AI correction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedacaoInput"
      responses:
        "201":
          description: Submitted essay with AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redacao"
  /redacoes/{id}:
    get:
      operationId: getRedacao
      tags: [redacao]
      summary: Get essay with AI feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Essay detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redacao"

  # ── Rankings ───────────────────────────────────────────────────────────────
...
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      summary: Mark a mission as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
components:
  schemas:
...
    FlashcardDeck:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        cardCount:
          type: integer
        dueCount:
          type: integer
        masteredCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, subject, cardCount, dueCount, masteredCount, createdAt]
    FlashcardDeckInput:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
      required: [title, subject]
    Flashcard:
      type: object
      properties:
        id:
          type: integer
        deckId:
          type: integer
        front:
          type: string
        back:
          type: string
        easeFactor:
          type: number
        interval:
          type: integer
        repetitions:
          type: integer
        nextReviewAt:
          type: ["string", "null"]
        mastered:
          type: boolean
      required: [id, deckId, front, back, easeFactor, interval, repetitions, mastered]
    FlashcardInput:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
      required: [front, back]
    FlashcardReview:
      type: object
      properties:
        quality:
          type: integer
          description: 0-5 SM2 quality rating
      required: [quality]
...
    Redacao:
      type: object
      properties:
        id:
          type: integer
        theme:
          type: string
        content:
          type: string
        score:
          type: ["integer", "null"]
        maxScore:
          type: integer
        feedback:
          type: ["string", "null"]
        competencias:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/CompetenciaScore"
        createdAt:
          type: string
      required: [id, theme, content, maxScore, createdAt]
    RedacaoInput:
      type: object
      properties:
        theme:
          type: string
        content:
          type: string
      required: [theme, content]
    CompetenciaScore:
      type: object
      properties:
        number:
          type: integer
        description:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        feedback:
          type: string
      required: [number, description, score, maxScore, feedback]
...
    Mission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        xpReward:
          type: integer
        coinReward:
          type: integer
        completed:
          type: boolean
        progress:
          type: integer
        target:
          type: integer
        type:
          type: string
          description: study, flashcard, simulado, redacao
      required: [id, title, description, icon, xpReward, coinReward, completed, progress, target, type]