# lib/api-zod/src/generated/types/competenciaScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompetenciaScore {
  number: number;
  description: string;
  score: number;
  maxScore: number;
  feedback: string;
}