# lib/api-zod/src/generated/types/flashcard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Flashcard {
  id: number;
  deckId: number;
  front: string;
  back: string;
  easeFactor: number;
  interval: number;
  repetitions: number;
  /** @nullable */
  nextReviewAt?: string | null;
  mastered: boolean;
}