# lib/api-zod/src/generated/types/flashcardDeckInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlashcardDeckInput {
  title: string;
  subject: string;
}