# lib/api-zod/src/generated/types/flashcardReview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlashcardReview {
  /** 0-5 SM2 quality rating */
  quality: number;
}