# lib/api-zod/src/generated/types/redacao.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompetenciaScore } from './competenciaScore';

export interface Redacao {
  id: number;
  theme: string;
  content: string;
  /** @nullable */
  score?: number | null;
  maxScore: number;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  competencias?: CompetenciaScore[] | null;
  createdAt: string;
}